Add event UNIT_AURA as an alternative to PLAYER_CONTROL_GAINED
PLAYER_CONTROL_GAINED may not be fired when a flight ends. UNIT_AURA also fires due to "flight" just being the application of a hidden movement aura and loss of player control. So, it is sufficient to check for this event and add a precautionary check for unit speed below an apparent minimum flight unit speed of "32".